Let transcoder::totalsize be uint64_t

+++ src/transcode.c
@@ -514,7 +514,7 @@ int transcoder_read(struct transcoder* trans, char* buff, int offset, int len) {
      */
     if (offset > trans->buffer.pos
         && offset + len > (trans->totalsize - 128)) {
-        int id3start = trans->totalsize - 128;
+        uint64_t id3start = trans->totalsize - 128;
 
         // zero the buffer
         memset(buff, 0, len);
@@ -580,7 +580,7 @@ int transcoder_finish(struct transcoder* trans) {
         if (trans->buffer.pos + 128 != trans->totalsize) {
             // write the id3v1 tag, always 128 bytes from end
             mp3fs_debug("Something went wrong with file size calculation: "
-                  "off by %lu", trans->buffer.pos + 128 - trans->totalsize);
+                  "off by %jd", trans->buffer.pos + 128 - trans->totalsize);
             trans->buffer.pos = trans->totalsize - 128;
         }
         buffer_write(&trans->buffer, (uint8_t*)trans->id3v1tag, 128);
